// enums/error-code.ts
export enum ErrorCode {
  SOMETHING_WENT_WRONG = 'SOMETHING_WENT_WRONG',
  CANNOT_PARSE_REQUEST_BODY = 'CANNOT_PARSE_REQUEST_BODY',
  INVALID_DATA_VALIDATION = 'INVALID_DATA_VALIDATION',
  CANNOT_CREATE_REQUEST_TOKEN = 'CANNOT_CREATE_REQUEST_TOKEN',
  CANNOT_CREATE_SESSION_WITH_LOGIN = 'CANNOT_CREATE_SESSION_WITH_LOGIN',
  CANNOT_CREATE_SESSION = 'CANNOT_CREATE_SESSION',

  CANNOT_ADD_TO_WATCHLIST = 'CANNOT_ADD_TO_WATCHLIST',
  CANNOT_DELETE_FROM_WATCHLIST = 'CANNOT_DELETE_FROM_WATCHLIST',

  CANNOT_GET_FAVORITE_MOVIES = 'CANNOT_GET_FAVORITE_MOVIES',
  CANNOT_ADD_FAVORITE_MOVIE = 'CANNOT_ADD_FAVORITE_MOVIE',
  CANNOT_DELETE_FAVORITE_MOVIE = 'CANNOT_DELETE_FAVORITE_MOVIE',
  CANNOT_GET_MOVIES_WATCHLIST = 'CANNOT_GET_MOVIES_WATCHLIST',
  CANNOT_GET_RATED_MOVIES = 'CANNOT_GET_RATED_MOVIES',
  CANNOT_ADD_MOVIE_RATING = 'CANNOT_ADD_MOVIE_RATING',
  CANNOT_DELETE_MOVIE_RATING = 'CANNOT_DELETE_MOVIE_RATING',
  CANNOT_GET_MOVIES_SEARCH = 'CANNOT_GET_MOVIES_SEARCH',
  CANNOT_GET_MOVIES_NOW_PLAYING = 'CANNOT_GET_MOVIES_NOW_PLAYING',
  CANNOT_GET_MOVIES_POPULAR = 'CANNOT_GET_MOVIES_POPULAR',
  CANNOT_GET_MOVIES_UPCOMING = 'CANNOT_GET_MOVIES_UPCOMING',
  CANNOT_GET_MOVIES_TRENDING = 'CANNOT_GET_MOVIES_TRENDING',
  CANNOT_GET_MOVIES_GENRES = 'CANNOT_GET_MOVIES_GENRES',

  CANNOT_GET_FAVORITE_TV_SHOWS = 'CANNOT_GET_FAVORITE_TV_SHOWS',
  CANNOT_ADD_FAVORITE_TV_SHOW = 'CANNOT_ADD_FAVORITE_TV_SHOW',
  CANNOT_DELETE_FAVORITE_TV_SHOW = 'CANNOT_DELETE_FAVORITE_TV_SHOW',
  CANNOT_GET_TV_SHOWS_WATCHLIST = 'CANNOT_GET_TV_SHOWS_WATCHLIST',
  CANNOT_GET_TV_SHOWS_RATED = 'CANNOT_GET_TV_SHOWS_RATED',
  CANNOT_ADD_TV_SHOW_RATING = 'CANNOT_ADD_TV_SHOW_RATING',
  CANNOT_DELETE_TV_SHOW_RATING = 'CANNOT_DELETE_TV_SHOW_RATING',
  CANNOT_GET_TV_SHOWS_SEARCH = 'CANNOT_GET_TV_SHOWS_SEARCH',
  CANNOT_GET_TV_SHOWS_AIRING_TODAY = 'CANNOT_GET_TV_SHOWS_AIRING_TODAY',
  CANNOT_GET_TV_SHOWS_POPULAR = 'CANNOT_GET_TV_SHOWS_POPULAR',
  CANNOT_GET_TV_SHOWS_ON_THE_AIR = 'CANNOT_GET_TV_SHOWS_ON_THE_AIR',
  CANNOT_GET_TV_SHOWS_TRENDING = 'CANNOT_GET_TV_SHOWS_TRENDING',
  CANNOT_GET_TV_SHOWS_GENRES = 'CANNOT_GET_TV_SHOWS_GENRES',
}