// app/api/movies-and-tv-shows/movies/watchlist/types.ts
import { Language } from '@/enums/language';
import { z } from 'zod';

export const getMoviesWatchlistRequestSchema = z.object({
  session_id: z.string(),
  language: z.nativeEnum(Language).optional(),
  page: z.number().int().optional(),
  sort_by: z.enum(['created_at.asc', 'created_at.desc']).optional(),
});

export type GetMoviesWatchlistRequest = z.infer<typeof getMoviesWatchlistRequestSchema>;